Makes Config module for ipaddr, port, etc.
Prior to this commit we were not exposing the two reference variables
ipaddr & port which resulted in a missing value error when sending a
metric. Furthermore, we expose the logger functions `log_debug` and
`log_error` as these were exposed before these separate package
commits.
In order to preserve backwards compatibility for the `statsd-client`
package with respect to these configuration variables, I had to symlink
to the `Statsd_client_lwt` module in the `statsd-client`
directory. Without doing this, setting `Statsd_client.{ipaddr,port}`
variables would work for `Statsd_client.Sync` functions but
`Statsd_client.Lwt` functions would throw an error that these variables
are undefined. By making this symlink in the `statsd-client` directory,
the `Statsd_client_lwt` module is built with the other modules and uses
the expected variables (as opposed to being a dependent package).
In the future, I'd like to get rid of this symlink behavior but this
requires a breaking change which at the moment we are avoiding for
the statsd-client package.